Initial commit
A simplistic command line tool that reads UW1's level.ark from player's save, and then dumps some information from it as a text.
This is primarily to be able to search the item locations inside the level.

Picked out number of utils from AGS source code. Hopefully this is permitted to post this under a MIT license because 95% of the code was mine, and the rest is trivial... I can always fix license notes later.